# crypto/merkle: hashing ~10% faster and ~60% fewer allocs