--- a/crypto/merkle/proof_test.go
+++ b/crypto/merkle/proof_test.go
@@ -171,7 +171,6 @@ func TestProofValidateBasic(t *testing.T) {
 	}
 }
 func TestVoteProtobuf(t *testing.T) {
-
 	_, proofs := ProofsFromByteSlices([][]byte{
 		[]byte("apple"),
 		[]byte("watermelon"),