This PR makes some optimizations that produce the following results compared to the `master` implementation:
```
name                          old time/op    new time/op    delta
HashAlternatives/recursive-8    72.2µs ± 6%    65.0µs ± 1%   -9.94%  (p=0.000 n=10+10)
HashAlternatives/iterative-8    71.5µs ± 1%    65.6µs ± 1%   -8.26%  (p=0.000 n=8+8)

name                          old alloc/op   new alloc/op   delta
HashAlternatives/recursive-8    25.4kB ± 0%     6.5kB ± 0%  -74.45%  (p=0.000 n=10+10)
HashAlternatives/iterative-8    28.1kB ± 0%     9.2kB ± 0%  -67.33%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
HashAlternatives/recursive-8       497 ± 0%       200 ± 0%  -59.76%  (p=0.000 n=10+10)
HashAlternatives/iterative-8       498 ± 0%       201 ± 0%  -59.64%  (p=0.000 n=10+10)
```

The gist of the change is basically avoiding allocs, and some nit gain by using a SIMD sha256 implementation. The latter could also be considered for the `crypto/tmhash` package.

There's a further optimization possible that might reduce allocs even more with pooling, but the main bottleneck is sha256 calc as expected so maybe isn't worth it.

I kept the original `leftHash` and `innerHash` methods since they're used in other places, but looks to me can be removed and used the ones I introduced which will avoid the same kind in allocations in other places producing some performance gains in other packages.