@jsign 2021-05-22
Looking now at `crytpo/merkle/proof.go`, definitely looks like a good idea to use the `*Opt` versions introduced here that would result in a similar gain regarding avoiding allocs for `ProofsFromByteSlices`. 

I didn't want to cascade the change to more places until having some extra feedback of this feels correct. I can extend this PR to only remove the previous functions if that sounds ok.