@liamsi 2021-05-23
Wow, thanks for doing this (ref #272) and for providing proper benchmarks! I'll do a proper review shortly.