@jsign 2021-05-24
@liamsi, I prefer going the route of leaving SIMD out of this PR as to prove the impact of reducing allocs.

As to be totally sure about the benchstat result _without_ SIMD, I've run it again:
```
benchstat master.bench opt.bench                                       
name                          old time/op    new time/op    delta
HashAlternatives/recursive-8    75.4µs ± 7%    66.5µs ± 2%  -11.85%  (p=0.000 n=10+9)
HashAlternatives/iterative-8    76.3µs ± 4%    64.7µs ± 1%  -15.18%  (p=0.000 n=10+9)

name                          old alloc/op   new alloc/op   delta
HashAlternatives/recursive-8    25.4kB ± 0%     6.5kB ± 0%  -74.45%  (p=0.000 n=10+10)
HashAlternatives/iterative-8    28.1kB ± 0%     9.2kB ± 0%  -67.33%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
HashAlternatives/recursive-8       497 ± 0%       200 ± 0%  -59.76%  (p=0.000 n=10+10)
HashAlternatives/iterative-8       498 ± 0%       201 ± 0%  -59.64%  (p=0.000 n=10+10)
```
My laptop is a bit noisy, but this proves that the gains are mostly from reducing allocs.

The test data is also quite tiny as to have much impact from SIMD instructions, so makes sense.

I agree on leaving the dependency change to another PR.